+++ Runtime/OnnxResizeAlgorithm.cs
@@ -1,0 +1,8 @@
+namespace OnnxRuntimeInference
+{
+    public enum OnnxResizeAlgorithm
+    {
+        Nearest = 0,
+        Bilinear = 1
+    }
+}